Python for GCP/Infra_with_python/requirements.txt
google-api-python-client
google-cloud-storage
